+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add `outgoing_trust_ukprn` to Transfer::Project model
+
 ### Changed
 
 - All `regional-casework-services` URLs are now `team`, e.g.